remove unused packages
refactor create-solid code
