--- packages/commands/.gitignore
@@ -1,1 +1,0 @@
-types
