--- packages/commands/package.json
@@ -1,53 +1,0 @@
-{
-	"name": "@solid-cli/commands",
-	"version": "0.0.24",
-	"description": "The main command handlers for the Solid CLI",
-	"license": "MIT",
-	"homepage": "https://solid-cli.netlify.app",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/solidjs-community/solid-cli"
-	},
-	"files": [
-		"dist",
-		"types"
-	],
-	"main": "./dist/index.mjs",
-	"module": "./dist/index.mjs",
-	"types": "./types/index.d.ts",
-	"exports": {
-		".": {
-			"import": "./dist/index.mjs",
-			"require": "./dist/index.mjs",
-			"types": "./types/index.d.ts"
-		},
-		"./new": {
-			"import": "./dist/handlers/new.mjs",
-			"require": "./dist/handlers/new.mjs",
-			"types": "./types/handlers/new.d.ts"
-		}
-	},
-	"scripts": {
-		"test": "vitest run",
-		"build": "tsc && tsup"
-	},
-	"devDependencies": {
-		"@chialab/esbuild-plugin-meta-url": "^0.18.2",
-		"@types/node": "^22.9.0",
-		"prettier": "^3.3.3",
-		"tsup": "^8.3.5",
-		"typescript": "^5.6.3"
-	},
-	"publishConfig": {
-		"access": "public"
-	},
-	"dependencies": {
-		"@begit/core": "^0.0.19",
-		"@clack/prompts": "0.7.0",
-		"@solid-cli/ui": "workspace:*",
-		"@solid-cli/utils": "workspace:*",
-		"execa": "^8.0.1",
-		"picocolors": "^1.1.1",
-		"sucrase": "^3.35.0"
-	}
-}
