--- packages/commands/src/handlers/start/index.ts
@@ -1,5 +1,0 @@
-export * from "./adapter";
-export * from "./api";
-export * from "./data";
-export * from "./mode";
-export * from "./route";
