--- packages/commands/src/index.ts
@@ -1,3 +1,0 @@
-export * from "./handlers/add";
-export * from "./handlers/new";
-export * from "./handlers/start";
