--- packages/commands/tests/assets/sample_app_config.txt
@@ -1,7 +1,0 @@
-import { defineConfig } from "@solidjs/start/config";
-// Simulates an app config
-export default defineConfig({
-  vite: {
-    plugins: [solid()]
-  }
-});
