--- packages/commands/tests/assets/sample_unocss_app_result.txt
@@ -1,11 +1,0 @@
-import UnoCss from "unocss/vite";
-import { defineConfig } from "@solidjs/start/config";
-// Simulates an app config
-export default defineConfig({
-  vite: {
-    plugins: [
-        solid(),
-        UnoCss({})
-    ]
-  }
-});
