--- packages/commands/tests/assets/sample_unocss_vite_result.txt
@@ -1,9 +1,0 @@
-import UnoCss from "unocss/vite";
-import { defineConfig } from "vite";
-// Simulates an app config
-export default defineConfig({
-  plugins: [
-      solid(),
-      UnoCss({})
-  ]
-});
