--- packages/commands/tests/assets/sample_vite_config.txt
@@ -1,5 +1,0 @@
-import { defineConfig } from "vite";
-// Simulates an app config
-export default defineConfig({
-  plugins: [solid()]
-});
