--- packages/commands/tests/package_installs.test.ts
@@ -1,7 +1,0 @@
-import { expect, describe, it } from "vitest";
-
-describe("Package Installs", () => {
-	it("should install packages correctly", () => {
-		expect(true).toBe(true);
-	});
-});
