--- packages/commands/tsconfig.json
@@ -1,20 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "ESNext",
-		"useDefineForClassFields": true,
-		"module": "ESNext",
-		"skipLibCheck": true,
-		"declaration": true,
-		"emitDeclarationOnly": true,
-		"moduleResolution": "Bundler",
-		"resolveJsonModule": true,
-		"isolatedModules": true,
-		"allowSyntheticDefaultImports": true,
-		"strict": true,
-		// "noUnusedLocals": true,
-		// "noUnusedParameters": true,
-		"noFallthroughCasesInSwitch": true,
-		"outDir": "types"
-	},
-	"include": ["src/*"]
-}
