--- packages/core/README.md
@@ -1,48 +1,0 @@
-# @solid-cli/core
-
-The main package for the SolidJS CLI tool.
-
-## Installation
-
-Run the following command to install the CLI tool globally:
-
-```bash
-npm i -g @solid-cli/core
-```
-
-## Usage
-
-For a full list of commands and how to use them, you can invoke the help command like so:
-
-```bash
-solid --help
-```
-
-Which should print this:
-
-```bash
-solid <subcommand>
-
-where <subcommand> can be one of:
-
-- add - Can add and install integrations: `solid add unocss`.
-- docs
-- new - Creates a new solid project
-- set
-- start - Commands specific to solid start
-- playground
-
-For more help, try running `solid <subcommand> --help`
-```
-
-## Documentation
-
-Visit our [official documentation](https://solid-cli.netlify.app/)
-
-## Support
-
-Reach out to us on the official [SolidJS discord](https://discord.gg/solidjs)
-
-## Contributing
-
-If you would like to contribute or keep up to date with the package, come join the effort over on [our GitHub](https://github.com/solidjs-community/solid-cli)
