--- packages/core/package.json
@@ -1,50 +1,0 @@
-{
-	"name": "@solid-cli/core",
-	"version": "0.0.26",
-	"description": "A CLI for Solid",
-	"author": "Thomas Beer & Rahul Batra",
-	"license": "MIT",
-	"homepage": "https://solid-cli.netlify.app",
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/solidjs-community/solid-cli"
-	},
-	"keywords": [
-		"solidjs",
-		"solid",
-		"cli"
-	],
-	"main": "dist/index.mjs",
-	"bin": {
-		"solid": "./dist/index.mjs"
-	},
-	"files": [
-		"dist"
-	],
-	"dependencies": {
-		"@clack/core": "0.3.4",
-		"@clack/prompts": "0.7.0",
-		"@solid-cli/commands": "workspace:*",
-		"@solid-cli/ui": "workspace:*",
-		"@solid-cli/utils": "workspace:*",
-		"cmd-ts": "^0.13.0",
-		"execa": "^8.0.1",
-		"picocolors": "^1.1.1",
-		"smol-toml": "^1.3.0",
-		"tiny-updater": "^3.5.2"
-	},
-	"scripts": {
-		"start": "jiti ./src/index.ts",
-		"build": "tsc && tsup"
-	},
-	"devDependencies": {
-		"@chialab/esbuild-plugin-meta-url": "^0.18.2",
-		"@types/node": "^22.9.0",
-		"jiti": "^2.4.0",
-		"tsup": "^8.3.5",
-		"typescript": "^5.6.3"
-	},
-	"publishConfig": {
-		"access": "public"
-	}
-}
