--- packages/core/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"compilerOptions": {
-		"baseUrl": ".",
-		"noEmit": true,
-		"module": "ESNext",
-		"target": "ESNext",
-		"moduleResolution": "Bundler",
-		"strict": true,
-		"esModuleInterop": true,
-		"forceConsistentCasingInFileNames": true,
-		"skipLibCheck": true,
-		"resolveJsonModule": true
-	}
-}
