--- packages/core/tsup.config.ts
@@ -1,13 +1,0 @@
-import { defineConfig } from "tsup";
-import metaUrlPlugin from "@chialab/esbuild-plugin-meta-url";
-
-export default defineConfig({
-	entry: ["src/index.ts"],
-	target: "esnext",
-	format: "esm",
-	splitting: false,
-	sourcemap: true,
-	clean: true,
-	minify: true,
-	esbuildPlugins: [metaUrlPlugin()],
-});
