+++ packages/create-solid/src/index.ts
@@ -5,9 +5,10 @@ import packageJson from "../package.json" with { type: "json" };
 import { defineCommand, runMain } from "citty";
 import { createVanilla, createVanillaJS } from "./create-vanilla";
 import * as p from "@clack/prompts";
-import { cancelable, getTemplatesList, spinnerify, StartTemplate, VanillaTemplate } from "./helpers";
+import { cancelable, spinnerify } from "./utils/ui";
 import { createStart } from "./create-start";
 import { create } from "./create";
+import { getTemplatesList, StartTemplate, VanillaTemplate } from "./utils/constants";
 intro(`\n${color.bgCyan(color.black(` Create-Solid v${packageJson.version}`))}`);
 
 const main = defineCommand({
