+++ packages/create-solid/src/utils/ui.ts
@@ -1,34 +1,3 @@
-export const GIT_IGNORE = `
-dist
-.solid
-.output
-.vercel
-.netlify
-.vinxi
-app.config.timestamp_*.js
-
-# Environment
-.env
-.env*.local
-
-# dependencies
-/node_modules
-
-# IDEs and editors
-/.idea
-.project
-.classpath
-*.launch
-.settings/
-
-# Temp
-gitignore
-
-# System Files
-.DS_Store
-Thumbs.db
-`;
-
 import { log, spinner } from "@clack/prompts";
 type SpinnerItem<T> = {
     startText: string;
@@ -59,14 +28,3 @@ export const cancelable = async <T = unknown>(prompt: Promise<T | symbol>, cance
 
     return value;
 };
-const VANILLA_TEMPLATES = ["ts"] as const;
-export type VanillaTemplate = (typeof VANILLA_TEMPLATES)[number];
-
-const START_TEMPLATES = ["basic"] as const;
-export type StartTemplate = (typeof VANILLA_TEMPLATES)[number];
-export const getTemplatesList = async (isStart: boolean) => {
-    if (isStart) {
-        return START_TEMPLATES;
-    }
-    return VANILLA_TEMPLATES;
-}
